+++ src/utils/token/classify/classifyToken.ts
@@ -1,26 +1,6 @@
 import { z } from "zod";
+import { Asset, FA12Token, FA2Token, NFT } from "../../../types/Asset";
 import { Token } from "../../../types/Token";
-
-type TokenBase = {
-  contract: string;
-  balance: string;
-};
-
-type FA2TokenMetadata = {
-  name?: string;
-  symbol?: string;
-  decimals?: string;
-};
-
-type NFTMetadata = {
-  name?: string;
-  symbol?: string;
-  displayUri: string;
-};
-
-export type FA12Token = TokenBase;
-export type FA2Token = TokenBase & { id: number; metadata: FA2TokenMetadata };
-export type NFT = TokenBase & { id: number; metadata: NFTMetadata };
 
 /**
  * Runtime validations with zod
@@ -136,5 +116,5 @@ const makeNft = (json: Token): NFT | null => {
   };
 };
 
-export const classifyToken = (json: Token) =>
+export const classifyToken = (json: Token): Asset | null =>
   makeFa1(json) || makeNft(json) || makeFa2(json);
